pyproject.toml
[tool.poetry]
name = "cybercat"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.3"
matplotlib = "^3.8.2"
scipy = "^1.11.4"
pygame = "^2.5.2"
pyaudio = "^0.2.14"
adafruit-circuitpython-neopixel = "^6.3.11"
rpi-gpio = "^0.7.1"
rpi-ws281x = "^5.0.0"
nicegui = "^1.4.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
